Lots of stuff...

+++ scripts/prepare-machine.ps1
@@ -63,7 +63,10 @@ param (
     [switch]$DuoNic,
 
     [Parameter(Mandatory = $false)]
-    [switch]$NoCodeCoverage
+    [switch]$NoCodeCoverage,
+
+    [Parameter(Mandatory = $false)]
+    [switch]$Xdp
 )
 
 #Requires -RunAsAdministrator
@@ -129,6 +132,18 @@ function Download-CoreNet-Deps {
         $ZipPath = Join-Path $ArtifactsPath "corenet-ci.zip"
         Invoke-WebRequest -Uri "https://github.com/microsoft/corenet-ci/archive/refs/heads/main.zip" -OutFile $ZipPath
         Expand-Archive -Path $ZipPath -DestinationPath $ArtifactsPath -Force
+        Remove-Item -Path $ZipPath
+    }
+}
+
+function Download-Xdp-Kit {
+    if (!(Test-Path $ArtifactsPath)) { mkdir $ArtifactsPath }
+    $XdpPath = Join-Path $ArtifactsPath "xdp"
+    if (!(Test-Path $XdpPath)) {
+        Write-Host "Downloading XDP Kit"
+        $ZipPath = Join-Path $ArtifactsPath "xdp.zip"
+        Invoke-WebRequest -Uri "https://lolafiles.blob.core.windows.net/nibanks/xdp.zip" -OutFile $ZipPath
+        Expand-Archive -Path $ZipPath -DestinationPath $XdpPath -Force
         Remove-Item -Path $ZipPath
     }
 }
@@ -321,6 +336,10 @@ if ($IsWindows) {
         if ($DuoNic) {
             Install-DuoNic
         }
+    }
+
+    if ($Xdp) {
+        Download-Xdp-Kit
     }
 
 } elseif ($IsLinux) {
